Add EMPTY_ARR to common constants, small refactor
