+++ src/components/Icon/hooks/useIconProps.tsx
@@ -1,9 +1,9 @@
-import { KeyboardEvent, UIEvent, MouseEvent, useCallback, useMemo, useRef } from "react";
+import { KeyboardEvent, MouseEvent, UIEvent, useCallback, useMemo, useRef } from "react";
 import cx from "classnames";
 import NOOP from "lodash/noop";
 import useEventListener from "../../../hooks/useEventListener";
 import useKeyEvent from "../../../hooks/useKeyEvent";
-import { keyCodes } from "../../../constants/KeyCodes";
+import { keyCodes } from "../../../constants/keyCodes";
 import useIconScreenReaderAccessProps from "../../../hooks/useIconScreenReaderAccessProps";
 
 const KEYS = [keyCodes.ENTER, keyCodes.SPACE];
