+++ src/constants/common.ts
@@ -1,0 +1,1 @@
+export const EMPTY_ARR: any[] = [];
