+++ src/constants/index.js
@@ -1,4 +1,5 @@
 export * from "./dialog";
-export * from "./KeyCodes";
+export * from "./keyCodes";
 export * from "./sizes";
 export * from "./positions";
+export * from "./common";
